fix: revalidateRouteHandler path for App Router (#896)
The `revalidateRouteHandler` function uses the path pattern for the
pages router: `/_sites/${host}${path}`. When
`revalidatePath(pathToRevalidate)` is called, the `pathToRevalidate` is
incorrect, resulting in WordPress content updates not appearing on the
frontend without a manual cache purge.

### How to test the Change
On a site with caching enabled, update a WordPress page. The updated
content will not automatically appear on the frontend (when using App
Router). The `/api/revalidate` endpoint is called, but the computed path
to revalidate is incorrect.

### Changelog Entry
> Changed - With this change, the correct path is created, resulting in
a properly revalidated path.
> Added - A new optional callback parameter was added to
`revalidateRouteHandler`. This callback allows extra logic to be run
after the path is revalidated. For instance, revalidateTag may need to
be called for a specific query, or a `redis` tag may need to be deleted.

### Checklist:
<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you are unsure about any of these, please ask for
clarification. We are here to help! -->
- [x] I agree to follow this project's [**Code of
Conduct**](https://github.com/10up/.github/blob/trunk/CODE_OF_CONDUCT.md).
- [ ] I have updated the documentation accordingly.
- [ ] I have added [Critical Flows, Test Cases, and/or End-to-End
Tests](https://10up.github.io/Open-Source-Best-Practices/testing/) to
cover my change.
- [x] All new and existing tests pass.